// src/types/thirdparty.type.ts
export interface CoinGeckoToken {
    id: string;
    symbol: string;
    name: string;
    image: string;
    current_price: number;
    market_cap: number;
    market_cap_rank: number;
    total_volume: number;
    price_change_percentage_24h: number;
    sparkline_in_7d: { price: number[] };
    categories: string[];
}